# swagger/openemr-api.yaml
openapi: 3.0.0
info:
  title: 'OpenEMR API'
  version: 7.0.0
servers:
  -
    url: /apis/default/
paths:
...
  '/api/patient/{puuid}/encounter':
    get:
      tags:
        - standard
      description: 'Retrieves a list of encounters for a single patient'
      parameters:
        -
          name: puuid
          in: path
          description: 'The uuid for the patient.'
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/standard'
        '400':
          $ref: '#/components/responses/badrequest'
        '401':
          $ref: '#/components/responses/unauthorized'
      security:
        -
          openemr_auth: []
    post:
      tags:
        - standard
      description: 'Creates a new encounter'
      parameters:
        -
          name: puuid
          in: path
          description: 'The uuid for the patient.'
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api_encounter_request'
      responses:
        '200':
          description: 'Standard response'
          content:
            application/json:
              schema:
                properties:
                  validationErrors: { description: 'Validation errors.', type: array, items: { type: object } }
                  internalErrors: { description: 'Internal errors.', type: array, items: { type: object } }
                  data: { description: 'Returned data.', type: array, items: { properties: { encounter: { description: 'encounter id', type: integer }, uuid: { description: 'encounter uuid', type: string } }, type: object } }
                type: object
                example:
                  validationErrors: []
                  error_description: []
                  data: { encounter: 1, uuid: 90c196f2-51cc-4655-8858-3a80aebff3ef }
        '401':
          $ref: '#/components/responses/unauthorized'
      security:
        -
          openemr_auth: []
  '/api/patient/{puuid}/encounter/{euuid}':
    get:
      tags:
        - standard
      description: 'Retrieves a single encounter for a patient'
      parameters:
        -
          name: puuid
          in: path
          description: 'The uuid for the patient.'
          required: true
          schema:
            type: string
        -
          name: euuid
          in: path
          description: 'The uuid for the encounter.'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Standard response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api_encounter_response'
        '401':
          $ref: '#/components/responses/unauthorized'
      security:
        -
          openemr_auth: []
    put:
      tags:
        - standard
      description: 'Modify a encounter'
      parameters:
        -
          name: puuid
          in: path
          description: 'The uuid for the patient.'
          required: true
          schema:
            type: string
        -
          name: euuid
          in: path
          description: 'The uuid for the encounter.'
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api_encounter_request'
      responses:
        '200':
          description: 'Standard response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api_encounter_response'
        '401':
          $ref: '#/components/responses/unauthorized'
      security:
        -
          openemr_auth: []
...
  /fhir/Encounter:
    get:
      tags:
        - fhir
      description: 'Returns a list of Encounter resources.'
      parameters:
        -
          name: _id
          in: query
          description: 'The uuid for the Encounter resource.'
          required: false
          schema:
            type: string
        -
          name: patient
          in: query
          description: 'The uuid for the patient.'
          required: false
          schema:
            type: string
        -
          name: date
          in: query
          description: 'The datetime of the Encounter resource.'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'Standard Response'
          content:
            application/json:
              schema:
                properties:
                  'json object': { description: 'FHIR Json object.', type: object }
                type: object
                example:
                  meta: { lastUpdated: '2021-09-14T09:13:51' }
                  resourceType: Bundle
                  type: collection
                  total: 0
                  link: [{ relation: self, url: 'https://localhost:9300/apis/default/fhir/Encounter' }]
        '400':
          $ref: '#/components/responses/badrequest'
        '401':
          $ref: '#/components/responses/unauthorized'
      security:
        -
          openemr_auth: []
  '/fhir/Encounter/{uuid}':
    get:
      tags:
        - fhir
      description: 'Returns a single Encounter resource.'
      parameters:
        -
          name: uuid
          in: path
          description: 'The uuid for the Encounter resource.'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Standard Response'
          content:
            application/json:
              schema:
                properties:
                  'json object': { description: 'FHIR Json object.', type: object }
                type: object
                example:
                  id: 946da61d-6b95-4f8e-abe5-534a25913b71
                  meta: { versionId: '1', lastUpdated: '2021-09-19T06:27:41+00:00' }
                  resourceType: Encounter
                  identifier: [{ system: 'urn:ietf:rfc:3986', value: 946da61d-6b95-4f8e-abe5-534a25913b71 }]
                  status: finished
                  class: { system: 'http://terminology.hl7.org/CodeSystem/v3-ActCode', code: AMB, display: ambulatory }
                  type: [{ coding: [{ system: 'http://snomed.info/sct', code: '185349003', display: 'Encounter for check up (procedure)' }] }]
                  subject: { reference: Patient/946da61b-626b-4f88-81e2-adfb88f4f0fe, type: Patient }
                  participant: [{ type: [{ coding: [{ system: 'http://terminology.hl7.org/CodeSystem/v3-ParticipationType', code: PPRF, display: 'Primary Performer' }] }], period: { start: '2012-08-13T00:00:00+00:00' }, individual: { reference: Practitioner/946da61d-ac5f-4fdc-b3f2-7b58dc49976b, type: Practitioner } }]
                  period: { start: '2012-08-13T00:00:00+00:00' }
        '400':
          $ref: '#/components/responses/badrequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/uuidnotfound'
      security:
        -
          openemr_auth: []
...
  /portal/patient/encounter:
    get:
      tags:
        - standard-patient
      description: 'Returns encounters for the patient.'
      responses:
        '200':
          $ref: '#/components/responses/standard'
        '400':
          $ref: '#/components/responses/badrequest'
        '401':
          $ref: '#/components/responses/unauthorized'
      security:
        -
          openemr_auth: []
  '/portal/patient/encounter/{euuid}':
    get:
      tags:
        - standard-patient
      description: 'Returns a selected encounter by its uuid.'
      parameters:
        -
          name: euuid
          in: path
          description: 'The uuid for the encounter.'
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/standard'
        '400':
          $ref: '#/components/responses/badrequest'
        '401':
          $ref: '#/components/responses/unauthorized'
      security:
        -
          openemr_auth: []
...
components:
  schemas:
...
    api_encounter_request:
      description: 'Schema for the encounter request'
      required:
        - pc_catid
        - class_code
      properties:
        date:
          description: 'The date of encounter.'
          type: string
        onset_date:
          description: 'The onset date of encounter.'
          type: string
        reason:
          description: 'The reason of encounter.'
          type: string
        facility:
          description: 'The facility of encounter.'
          type: string
        pc_catid:
          description: 'The pc_catid of encounter.'
          type: string
        facility_id:
          description: 'The facility id of encounter.'
          type: string
        billing_facility:
          description: 'The billing facility id of encounter.'
          type: string
        sensitivity:
          description: 'The sensitivity of encounter.'
          type: string
        referral_source:
          description: 'The referral source of encounter.'
          type: string
        pos_code:
          description: 'The pos_code of encounter.'
          type: string
        external_id:
          description: 'The external id of encounter.'
          type: string
        provider_id:
          description: 'The provider id of encounter.'
          type: string
        class_code:
          description: 'The class_code of encounter.'
          type: string
      type: object
      example:
        date: '2020-11-10'
        onset_date: ''
        reason: 'Pregnancy Test'
        facility: 'Owerri General Hospital'
        pc_catid: '5'
        facility_id: '3'
        billing_facility: '3'
        sensitivity: normal
        referral_source: ''
        pos_code: '0'
        external_id: ''
        provider_id: '1'
        class_code: AMB
...
  responses:
    standard:
      description: 'Standard Response'
      content:
        application/json:
          schema:
            properties:
              validationErrors:
                description: 'Validation errors.'
                type: array
                items:
                  type: object
              internalErrors:
                description: 'Internal errors.'
                type: array
                items:
                  type: object
              data:
                description: 'Returned data.'
                type: array
                items:
                  type: object
            type: object
            example:
              validationErrors: []
              error_description: []
              data: []
    badrequest:
      description: 'Bad Request'
      content:
        application/json:
          schema:
            properties:
              validationErrors:
                description: 'Validation errors.'
                type: object
            type: object
            example:
              validationErrors:
                _id: 'The search field argument was invalid, improperly formatted, or could not be parsed.  Inner message: UUID columns must be a valid UUID string'
    unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            properties:
              error:
                description: 'The error.'
                type: string
              error_description:
                description: 'The description of the error.'
                type: string
              hint:
                description: 'More specific information on the error.'
                type: string
              message:
                description: 'Message regarding the error.'
                type: string
            type: object
            example:
              error: access_denied
              error_description: 'The resource owner or authorization server denied the request.'
              hint: 'Missing "Authorization" header'
              message: 'The resource owner or authorization server denied the request.'
    uuidnotfound:
      description: 'Not Found'
      content:
        application/json:
          schema:
            properties:
              empty:
                description: empty
                type: object
            type: object
            example: []
  securitySchemes:
    openemr_auth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/default/authorize
          tokenUrl: /oauth2/default/token
          refreshUrl: /oauth2/default/token
          scopes:
            openid: 'Generic mandatory scope'
            offline_access: 'Will signal server to provide a refresh token'
            launch/patient: 'Will provide a patient selector when logging in as an OpenEMR user (required for testing patient/* scopes in swagger if not logging in as a patient)'
            'api:fhir': 'FHIR R4 API'
            patient/AllergyIntolerance.read: 'Read allergy intolerance resources for the current patient (api:fhir)'
            patient/Appointment.read: 'Read appointment resources for the current patient (api:fhir)'
            patient/Binary.read: 'Read binary document resources for the current patient (api:fhir)'
            patient/CarePlan.read: 'Read care plan resources for the current patient (api:fhir)'
            patient/CareTeam.read: 'Read care team resources for the current patient (api:fhir)'
            patient/Condition.read: 'Read condition resources for the current patient (api:fhir)'
            patient/Coverage.read: 'Read coverage resources for the current patient (api:fhir)'
            patient/Device.read: 'Read device resources for the current patient (api:fhir)'
            patient/DiagnosticReport.read: 'Read diagnostic report resources for the current patient (api:fhir)'
            patient/DocumentReference.read: 'Read document reference resources for the current patient (api:fhir)'
            patient/DocumentReference.$docref: 'Generate a document for the current patient or returns the most current Clinical Summary of Care Document (CCD)'
            patient/Encounter.read: 'Read encounter resources for the current patient (api:fhir)'
            patient/Goal.read: 'Read goal resources for the current patient (api:fhir)'
            patient/Immunization.read: 'Read immunization resources for the current patient (api:fhir)'
            patient/Location.read: 'Read location resources for the current patient (api:fhir)'
            patient/Medication.read: 'Read medication resources for the current patient (api:fhir)'
            patient/MedicationRequest.read: 'Read medication request resources for the current patient (api:fhir)'
            patient/Observation.read: 'Read observation resources for the current patient (api:fhir)'
            patient/Organization.read: 'Read organization resources for the current patient (api:fhir)'
            patient/Patient.read: 'Read patient resource for the current patient (api:fhir)'
            patient/Person.read: 'Read person resources for the current patient (api:fhir)'
            patient/Practitioner.read: 'Read practitioner resources for the current patient (api:fhir)'
            patient/Procedure.read: 'Read procedure resources for the current patient (api:fhir)'
            patient/Provenance.read: 'Read provenance resources for the current patient (api:fhir)'
            system/AllergyIntolerance.read: 'Read all allergy intolerance resources in the system (api:fhir)'
            system/Binary.read: 'Read all binary document resources in the system (api:fhir)'
            system/CarePlan.read: 'Read all care plan resources in the system (api:fhir)'
            system/CareTeam.read: 'Read all care team resources in the system (api:fhir)'
...